feat(desktop)!: consolidate six shelf field tools into add-field / remove-field
Mirrors the a2td consolidation (a2td 6d5b958) in tmcp idiom: target
enum[rows,cols,encoding] with encodingType schema-optional and
handler-enforced required iff target=encoding; handlers switch(target)
and delegate unchanged to the existing metadata helpers. Six old
registrations deleted outright, no aliases — hosts hardcoding the old
names must move to the new ones.

Frees 3,000 serialized ListTools bytes (5,324 -> 2,324; both new tools
under the 1,200 per-tool budget) taking the surface to 42,818 of the
46,000 ToolSearch-deferral cliff — the headroom the refine-worksheet
port (claude/refine-route-port) is blocked on.
