+++ src/tools/desktop/toolName.ts
@@ -14,12 +14,8 @@ export const desktopToolNames = [
   'build-and-apply-dashboard',
   'list-available-fields',
   'list-fields',
-  'add-field-to-encoding',
-  'add-field-to-rows',
-  'add-field-to-cols',
-  'remove-field-from-encoding',
-  'remove-field-from-rows',
-  'remove-field-from-cols',
+  'add-field',
+  'remove-field',
   'resolve-field',
   'search-examples',
   'search-commands',
